[move/#203] 공통컴포넌트에서 record로 컴포넌트 이동
