+++ app/src/main/java/com/bongtu/baekseo/presentation/record/component/RecordScheduleList.kt
@@ -1,4 +1,4 @@
-package com.bongtu.baekseo.core.designsystem.component.list
+package com.bongtu.baekseo.presentation.record.component
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
@@ -25,7 +25,7 @@ import com.bongtu.baekseo.core.util.OnBottomReached
 import kotlinx.collections.immutable.ImmutableList
 
 @Composable
-fun <T> BongBaekScheduleList(
+fun <T> RecordScheduleList(
     items: ImmutableList<T>,
     getKey: (T) -> String,
     getDate: (T) -> String,
